requirements.txt
torch==1.8.1
pytorch-lightning==1.3.5
torchvision==0.9.1
PyYAML==5.4.1